+++ tests/nv/test_ex_routes.py
@@ -267,6 +267,7 @@ class TestRoutes(unittest.TestCase):
         self.assertRoute('_ex_route_history', ['history', 'his'])
         self.assertRoute('_ex_route_inoremap', ['inoremap', 'ino'])
         self.assertRoute('_ex_route_let', ['let '])
+        self.assertRoute('_ex_route_marks', ['marks'])
         self.assertRoute('_ex_route_move', ['move', 'm'])
         self.assertRoute('_ex_route_new', ['new'])
         self.assertRoute('_ex_route_nnoremap', ['nnoremap', 'nn'])